requirements.txt
aiohttp==3.6.2
altgraph==0.17
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
discord.py==1.3.4
idna==2.10
macholib==1.14
multidict==4.7.6
Pillow==7.2.0
PyInstaller==3.6
PySocks==1.7.1
requests==2.24.0
stem==1.8.0
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
